// src/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Decimal: any;
};

export type Asset = {
  __typename?: 'Asset';
  fireblocksTicker?: Maybe<Scalars['String']>;
  id?: Maybe<Scalars['Int']>;
  name: Scalars['String'];
  ticker?: Maybe<Scalars['String']>;
  type: Scalars['String'];
};
// ...
export type EligibleTradeAsset = {
  __typename?: 'EligibleTradeAsset';
  assets?: Maybe<Asset>;
  id: Scalars['Int'];
};
// ...
export type Order = {
  __typename?: 'Order';
  orderId: Scalars['String'];
};

export type OrderQueue = {
  __typename?: 'OrderQueue';
  id: Scalars['String'];
  orderPriceExecuted?: Maybe<Scalars['String']>;
  orderStatus: Scalars['String'];
  updatedAt: Scalars['DateTime'];
  userAcceptedPrice: Scalars['String'];
};
// ...
export type Quote = {
  __typename?: 'Quote';
  quoteId: Scalars['String'];
};

export type QuoteQueue = {
  __typename?: 'QuoteQueue';
  baseToken: Scalars['String'];
  finalQuoteValue?: Maybe<Scalars['String']>;
  id: Scalars['String'];
  quantity: Scalars['String'];
  quoteToken: Scalars['String'];
  side: Scalars['String'];
  spreadFee?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  userQuotePrice?: Maybe<Scalars['String']>;
};